+++ test/test_models.py
@@ -442,6 +442,30 @@ def test_observation__empty():
     assert obs.user is None
 
 
+def test_observation__application():
+    obs = Observation.from_json(j_observation_v2)
+    app = obs.application
+    assert isinstance(app, Application)
+    assert app.name == 'iNaturalist iPhone App'
+    assert app.url.startswith('https://itunes.apple.com')
+    assert (
+        str(app)
+        == 'Application(id=3, name=iNaturalist iPhone App, url=https://itunes.apple.com/us/app/inaturalist/id421397028?mt=8)'
+    )
+
+
+def test_observation__flags():
+    obs_json = deepcopy(j_observation_v2)
+    obs_json['flags'] = [j_flag_1]
+    obs = Observation.from_json(obs_json)
+    flag = obs.flags[0]
+    assert isinstance(flag, Flag)
+    assert flag.id == 123456
+    assert flag.resolved is False
+    assert flag.user.login == 'some_user'
+    assert str(flag) == 'Flag(id=123456, flag=spam, resolved=False, username=some_user)'
+
+
 def test_observation__ofvs():
     obs = Observation.from_json(j_observation_3_ofvs)
     ofv = obs.ofvs[0]
@@ -454,6 +478,30 @@ def test_observation__project_observations():
     obs = Observation.from_json(j_observation_3_ofvs)
     proj_obs = obs.project_observations[0]
     assert isinstance(proj_obs, ProjectObservation) and proj_obs.id == 48899479
+
+
+def test_observation__quality_metrics():
+    obs = Observation.from_json(j_observation_6_metrics)
+
+    metric = obs.quality_metrics[0]
+    assert isinstance(metric, QualityMetric)
+    assert metric.id == 6988064
+    assert metric.metric == 'wild'
+    assert metric.agree is True
+    assert metric.user.login == 'jkcook'
+    assert str(metric) == 'QualityMetric(id=6988064, metric=wild, agree=True, username=jkcook)'
+
+    vote = obs.votes[0]
+    assert isinstance(vote, Vote)
+    assert vote.vote_flag is True
+    assert vote.user.login == 'jkcook'
+    assert str(vote) == 'Vote(id=2115051, vote_flag=True, username=jkcook)'
+
+    fave = obs.faves[0]
+    assert isinstance(fave, Fave)
+    assert fave.vote_flag is True
+    assert fave.user.login == 'jkcook'
+    assert str(fave) == 'Fave(id=2115052, vote_flag=True, username=jkcook)'
 
 
 def test_observation__default_photo():